Removed recycled_copy.go, moved compareFilePaths to file_protection.go
